Makes test reusable
